% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{HPDint}
\alias{HPDint}
\title{HPD interval}
\usage{
HPDint(theta, prob = 0.95)
}
\arguments{
\item{theta}{vector of MCMC samples}

\item{prob}{probability (defaults to 0.95)}
}
\value{
parameter values for the lower and upper bound of the HPD
}
\description{
Finds the highest-posterior density (HPD) interval for the given probability.
}
