% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{HPDquantile}
\alias{HPDquantile}
\title{Quantiles of the HPD}
\usage{
HPDquantile(theta, outr = 0.95)
}
\arguments{
\item{theta}{vector of MCMC samples}

\item{outr}{outer range (default is 0.95)}
}
\value{
the vector that has the outer range as the first and last elements, the inter-quartile range as the second and penultimate elements, and the mode as the middle element.
}
\description{
Finds the mode, the inter-quartile range, and the outer range of the posterior distribution from an MCMC sample. The outer range must be > 0.5.
}
