% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{pmode}
\alias{pmode}
\title{Find posterior mode}
\usage{
pmode(theta)
}
\arguments{
\item{theta}{vector of MCMC samples of a variable}
}
\value{
value at the mode
}
\description{
Finds the posterior mode from an MCMC sample. If there are more than one, picks one randomly with a warning.
}
