% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSampler}
\alias{runSampler}
\title{Run MCMC}
\usage{
runSampler(nPos, nWells, dilFrac, nChains, nBurnin, nSample)
}
\arguments{
\item{nPos}{number of positive wells}

\item{nWells}{total number of wells for each dilution}

\item{dilFrac}{dilution fractions}

\item{nChains}{number of chains}

\item{nBurnin}{number of burn-in iterations}

\item{nSample}{number of sampling iterations}
}
\value{
A list containing chains of IUPM values, a vector of chain IDs, and a vector of acceptance rates
}
\description{
Fits a binomial model to well count data. Uses Metropolis-Hastings MCMC to generate samples from the posterior distribution of IUPM.
}
