% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{split.Rhat}
\alias{split.Rhat}
\title{Calculate rank-normalized \eqn{\widehat{R}}}
\usage{
\method{split}{Rhat}(theta, chnFac)
}
\arguments{
\item{theta}{vector of MCMC samples}

\item{chnFac}{factor marking chains}
}
\value{
maximum of the folded and unfolded \eqn{\widehat{R}}
}
\description{
Calculates the rank-normalized split-\eqn{\widehat{R}} statistic on regular and folded MCMC samples. Reports the maximum of the two. Folded \eqn{\widehat{R}} captures problems in exploring distribution tails. Values of \eqn{\widehat{R} < 0.01} are considered acceptable.
}
